fix(dtype): kernels and modules follow the checkpoint dtype, bf16 when undeclared
The dtype pipeline c6b5b36 built stopped at ModelConfig.dtype: everything
downstream still had fp16 hardcoded, so a bf16 checkpoint ran bf16 weights
through fp16-shaped decisions.

Kernels:
- fused_moe autotune keys follow the activation dtype (_QUANT_NONE was a
  hardcoded "fp16", so bf16 experts read fp16's tile table); quantised keys
  still name the weight format.
- fused_moe/w8a16 8-bit dequant casts straight to the activation's
  compute_type instead of hopping through tl.float16; e4m3 is exact in fp16
  and its 4 mantissa bits survive the bf16 hop, so this is a correctness of
  intent change, not a numeric one.
- w4a16: leftover merge markers in the docstring resolved (the fp32 zero-point
  subtraction they describe matches the implementation).

Modules/models:
- LinearBase and its four subclasses take a dtype argument (default bf16) and
  store it; quant methods read it back in create_weights, so an unquantised
  layer allocates in the checkpoint's element type instead of fp16.
- UnquantizedLinearMethod/FusedMoEMethod: get_name() "fp16" -> "unquantized",
  storage_dtype bf16, create_weights uses layer.dtype.
- SparseMoeBlock gate_weight and the vocab embedding / KV-cache-manager
  defaults follow config.dtype / bf16; base.py and mlp.py pass config.dtype
  into every projection they build.
- kv_cache_manager: indexs -> indices spellings (typos gate; identifiers are
  file-local).

Benchmarks: HF baselines load at the checkpoint's declared dtype
(checkpoint_dtype in common.py) so a bf16 model is not compared against an
fp16 HF row; fused_moe microbench rows re-baselined on bf16 with H100 numbers
(log: bench_fused_moe_h100_20260902.json).

Tests: the chunked-prefill engine tests compared 40 opening characters
byte-for-byte against the one-shot path. That is an fp16 property: under bf16
the flash-decoding extend route and the flash-attention grid route tile their
reductions differently enough to flip a near tie (measured top-2 logprob gap
0.125, 1.13:1). They now pin the first sampled token -- the signal a dropped
prefix actually moves -- and keep the ownership check. qwen3_moe router
reference aligns to the checkpoint dtype with a tolerance at bf16's rounding
floor; quant-method and weight-mapping fixtures assert the bf16 default.

Regression: kernels/ops/models/compile 562 passed; executor+kernels 261
after the rename; engine+golden 268 and compile 5 with a live checkpoint.

+++ lite_llama/kernels/ops/quantization/fp8.py
@@ -17,7 +17,7 @@ import triton.language as tl
 
 from .w8a16 import FP8_E4M3_BIT_TRICK_SCALE, dequant_fp8e4m3
 
-#: Exponent correction when *both* operands went through the e4m3 -> fp16 bit
+#: Exponent correction when *both* operands went through the e4m3 -> bf16 bit
 #: trick: each is short a factor of 256, so the product is short 256**2.
 _FP8_BIT_TRICK_SCALE_SQ = FP8_E4M3_BIT_TRICK_SCALE * FP8_E4M3_BIT_TRICK_SCALE
 
@@ -213,7 +213,7 @@ def _fp8_matmul_kernel(
         a = tl.load(a_ptrs, mask=offs_k[None, :] < k_rem, other=0)
         b = tl.load(b_ptrs, mask=offs_k[None, :] < k_rem, other=0)
         # NATIVE_FP8 is constexpr: Triton emits two specialised kernels, so
-        # the fp16 widening below never reaches an sm89+ binary (and the fp8
+        # the bf16 widening below never reaches an sm89+ binary (and the fp8
         # bitcast never reaches an sm86 one).
         if NATIVE_FP8:
             a = a.to(tl.float8e4nv, bitcast=True)
@@ -302,7 +302,7 @@ def fp8_matmul(
         group_n: Rows of one weight-scale block. ``1`` means per-output-channel.
         group_k: Columns of one weight-scale block. ``>= K`` means one per row.
         bias: Optional ``[N]`` bias, added in fp32 before the output cast.
-        out_dtype: Output dtype; the rest of the network runs bf16 (or fp16
+        out_dtype: Output dtype; the rest of the network runs bf16 (or bf16
             for a checkpoint that declares it).
 
     Returns:
